各種整理。
- build.gradle.ktsの導入。（ファイルリネーム。）
- MainからHouseholdAccountBookへのリネーム。
- 書き込むボタンの追加。
