+++ .idea/dictionaries/Tatsuro.xml
@@ -3,10 +3,12 @@
     <words>
       <w>androiddebugkey</w>
       <w>constraintlayout</w>
+      <w>householdaccountbook</w>
       <w>jetbrains</w>
       <w>kakei</w>
       <w>kakeinote</w>
       <w>kapt</w>
+      <w>kotlinx</w>
       <w>leakcanary</w>
       <w>livedata</w>
       <w>orhanobut</w>
