+++ app/src/main/java/com/tatsuro/app/kakeinote/ui/householdaccountbook/HouseholdAccountBookViewModel.kt
@@ -1,0 +1,8 @@
+package com.tatsuro.app.kakeinote.ui.householdaccountbook
+
+import androidx.lifecycle.ViewModel
+
+/** 家計簿ViewModel. */
+class HouseholdAccountBookViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
