--- app/src/main/java/com/tatsuro/app/kakeinote/ui/main/MainFragment.kt
@@ -1,14 +1,0 @@
-package com.tatsuro.app.kakeinote.ui.main
-
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import com.tatsuro.app.kakeinote.R
-
-class MainFragment : Fragment(R.layout.main_fragment) {
-
-    companion object {
-        fun newInstance() = MainFragment()
-    }
-
-    private val viewModel: MainViewModel by viewModels()
-}
