--- app/src/main/java/com/tatsuro/app/kakeinote/ui/main/MainViewModel.kt
@@ -1,7 +1,0 @@
-package com.tatsuro.app.kakeinote.ui.main
-
-import androidx.lifecycle.ViewModel
-
-class MainViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
