+++ app/src/main/res/layout/household_account_book_fragment.xml
@@ -2,9 +2,9 @@
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
+    android:id="@+id/mainLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.MainFragment">
+    tools:context=".ui.householdaccountbook.HouseholdAccountBookFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
