+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,5 @@
     <string name="app_name">かけいノート</string>
     <string name="app_name_short">かけい</string>
     <string name="household_account_book">家計簿</string>
+    <string name="write">書き込む</string>
 </resources>
