+++ app/src/main/res/values/themes.xml
@@ -1,16 +1,30 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools">
-    <!-- Base application theme. -->
+    <!-- アプリケーションのベーステーマ -->
     <style name="Theme.KakeiNote" parent="Theme.MaterialComponents.DayNight.NoActionBar">
-        <!-- Primary brand color. -->
+        <!-- プライマリの色 -->
         <item name="colorPrimary">@color/light_blue</item>
         <item name="colorPrimaryVariant">@color/light_blue</item>
         <item name="colorOnPrimary">@color/white</item>
-        <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/teal_200</item>
-        <item name="colorSecondaryVariant">@color/teal_700</item>
+        <!-- セカンダリの色 -->
+        <item name="colorSecondary">@color/light_blue</item>
+        <item name="colorSecondaryVariant">@color/light_blue</item>
         <item name="colorOnSecondary">@color/black</item>
-        <!-- Status bar color. -->
+        <!-- ステータスバーの色 -->
         <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
+    </style>
+
+    <!-- ツールバーのテーマ -->
+    <style name="Theme.KakeiNote.Toolbar">
+        <!-- ツールバーのタイトルテキストの色 -->
+        <item name="android:textColorPrimary">@color/white</item>
+        <!-- ツールバーのオプションメニューのOverflowアイコンの色 -->
+        <item name="android:textColorSecondary">@color/white</item>
+    </style>
+
+    <!-- FloatingActionButtonのテーマ -->
+    <style name="Theme.KakeiNote.FloatingActionButton">
+        <!-- FloatingActionButtonのテキストの色 -->
+        <item name="colorOnSecondary">@color/white</item>
     </style>
 </resources>
